Introduce a generic system for processing various data types.
A message / observation may be related to a value of any type, but it must keep track of a "data support", which is responsible of handling the nitty gritty associated with the type of the carried value : specifying column name in the cassandra cluster, encoding and decoding it from/to JSON and Cassandra, etc.

In the Kafka source, the type of a message must be specified using the "type" field. This breaks the format of the "raw" topic.

For the moment, only Double values are supported; an Kafka entry of any other type will result in an Deserialization exception.

The type support can be extended by

1. Implementing a DataAggregator for the supported data type;
2. Implementing the corresponding DataTypeSupport, which refers to the aggregator;
3. Adding the DataTypeSupport in Message.dataTypeSupports