Fix #45
It turns out the problem was not related to last double. Instead, adding extra aggregations causes a dead lock in the aggregation management loop, that eventually causes the entire process to stop.

In order to fix this problem, a buffer has been added right after the generated of the aggregated values (agg2agg), in such a way each aggregated values can become up to 5000 new aggregated values (which is considered as a very high limit). This buffer makes the whole process smoother.